fix styling of FileItemSimple by merging it with FileItem
