--- src/style/FileItemSimpleStyle.ts
@@ -1,9 +1,0 @@
-import { style } from 'typestyle';
-import { NestedCSSProperties } from 'typestyle/lib/types';
-
-export const gitMarkBoxCss: NestedCSSProperties = {
-  flex: '0 0 auto',
-  margin: 'auto 8px auto 4px'
-};
-
-export const gitMarkBoxStyle = style(gitMarkBoxCss);
